Corección de mapeo en firebase

+++ app/src/main/java/com/example/tp3_examen/components/CardBalance.kt
@@ -34,7 +34,7 @@ import androidx.compose.ui.res.stringResource
 @Composable
 fun CardBalance(
     cvu:String,
-    balance:Double
+    balance:Float
 ) {
     Card(
         colors = CardDefaults.cardColors(
