+++ app/src/main/java/com/example/tp3_examen/data/models/BankData.kt
@@ -2,6 +2,6 @@ package com.example.tp3_examen.data.models
 
 data class BankData(
     var cvu: String = "",
-    var balance: Double = 0.0,
+    var balance: Float = 0.0f,
     var transactions: List<BankAccountTransaction> = emptyList()
 )
